// src/serverApi/v3/models/user-match-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * HPI Schul-Cloud Server API
 * This is v3 of HPI Schul-Cloud Server. Checkout /docs for v1.
 *
 * The version of the OpenAPI document: 3.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface UserMatchResponse
 */
export interface UserMatchResponse {
    /**
     * local user id
     * @type {string}
     * @memberof UserMatchResponse
     */
    userId: string;
    /**
     * login name of local user
     * @type {string}
     * @memberof UserMatchResponse
     */
    loginName: string;
    /**
     * firstname of local user
     * @type {string}
     * @memberof UserMatchResponse
     */
    firstName: string;
    /**
     * lastname of local user
     * @type {string}
     * @memberof UserMatchResponse
     */
    lastName: string;
    /**
     * list of user roles from external system: student, teacher, admin
     * @type {Array<string>}
     * @memberof UserMatchResponse
     */
    roleNames: Array<UserMatchResponseRoleNames>;
    /**
     * match type: admin (manual) or auto (set, when names match exactly for a single user
     * @type {string}
     * @memberof UserMatchResponse
     */
    matchedBy?: UserMatchResponseMatchedBy;
}

/**
    * @export
    * @enum {string}
    */
export enum UserMatchResponseRoleNames {
    STUDENT = 'student',
    TEACHER = 'teacher',
    ADMIN = 'admin'
}
/**
    * @export
    * @enum {string}
    */
export enum UserMatchResponseMatchedBy {
    AUTO = 'auto',
    ADMIN = 'admin'
}